sign-detection-api/requirements.txt
Flask
numpy
Pillow
requests
tensorflow==2.12.1
Flask-CORS
Flask-SocketIO
mediapipe
opencv-python
pyngrok
uuid
ffmpeg-python